requirements.txt
requests
lxml
pushbullet.py

